feat: emit ChatSocket chat event when sending messages & update message list when receiving chat
