+++ frontend/src/components/Socket/SocketTest.jsx
@@ -1,34 +1,12 @@
 import React, { useEffect } from 'react';
-import io from 'socket.io-client';
 
-require('dotenv').config();
-
-const BASE_URL = process.env.REACT_APP_SOCKET_HOST;
-
-const createChatSocket = () => {
-    const chatSocket = io(`${BASE_URL}/chat`);
-
-    chatSocket.on('connect', () => {
-        // chatSocket.emit('join', { roomId: 'some random key' });
-        chatSocket.on('join', msg => console.log(msg));
-        // chatSocket.emit('chat', { message: 'Hello Server' });
-        chatSocket.on('chat', msg => console.log(msg));
-    });
-
-    chatSocket.on('disconnect', () => {
-        chatSocket.removeAllListeners();
-    });
-    return chatSocket;
-};
+import { ChatSocket } from '@/socket';
 
 const SocketTest = () => {
-    const chatSocket = createChatSocket();
     useEffect(() => {
         console.log('socket test');
-        chatSocket.on('connect', () => {
-            chatSocket.emit('join', { roomId: 'some random key' });
-            chatSocket.emit('chat', { message: 'Hello Server' });
-        });
+        ChatSocket.emit('join', { roomId: 'some random key' });
+        ChatSocket.emit('chat', { message: 'Hello Server' });
     }, []);
     return <></>;
 };
